Updated Reflection with other module tweaks to either accommodate the changes in Reflection or to fix issues that were apparent in the print statements.
For example, the Act module was being shown the
actions objects rather than the ACTION_DESCRIPTION.
I also needed to add another parameter to retrieve
to include IDs. I changed the character module to
include an include_perceptions parameter in the
get_standard_info method. Additionally, I commented
out the GPT description portion of the ok method in
the parsing class to avoid unnecessary calls to GPT
asking it to print/log a more elaborate description of
what happened. We can add the logging back later with
the basic action description. I don't think we need a
longer more detailed version.

Finally, I made a bunch of small tweaks to prompts that
needed spaces or newlines.